+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [CSE] Added remote resources support for &lt;group>.
 - [CSE] Added support for *dataGenerationTime* attribute for &lt;contentInstance> (R5 feature).
 - [CSE] Added "acme://" URL scheme for notifications to run ACMEScript scripts.
+- [CSE] Added always setting *Originating Timestamp* in requests and responses.
 - [WEB] Allow to open the WebUI of a registered CSE via the context menu.
 - [CONSOLE] Added config for dark (default) and light theme for better readability on consoles with light background.
 - [SCRIPTS] Added scripting to the CSE.